feat: 카페 리뷰 등록/수정 기능 변경
- 네트워킹 관련 기능 viewmodel로 이동
- 와이파이, 책상 필수 항목으로 변경
- LabeledRadioButton(삭제됨) -> 커스텀 UI (ReviewButtonGroup)으로 변경
- 리뷰 선택 해제 기능 추가
- RecyclerView 삭제

+++ app/src/main/java/com/example/mocacong/data/response/MyReviewResponse.kt
@@ -1,7 +1,7 @@
 package com.example.mocacong.data.response
 
 data class MyReviewResponse(
-    val myScore : Float,
+    val myScore : Int,
     val myStudyType: String?,
     val myWifi: String?,
     val myParking: String?,
